feat: Implemented the storage of the shows reviews in bigquery
This implementation is one of the finals steps of the pipeline and involves the population of a bigquery table that will serve as a source of truth for future analysis, also the show_id column was renamed in the consume_reddit_comments.py spark job. The final step should the validation of the data, this will be the next operator to be created.